Create readme in main directory
When GitHub finds a markdown-formatted readme file, it displays it on the front page of the project. Adding this file means that users accustomed to GitHub workflow are likelier to find the project's documentation when glancing at it, since this is the location where one expects to find that information.