Convert colors to given floss type before deciding to show colors from a square.
We take all of the colors from the original image in the square clicked on in
the squared image, and then when we're in a floss-restricted mode, convert
those colors to their corresponding floss-type colors.  If it turns out there's
only one of those floss-type colors then don't show the "square colors" option
in colorDialog (that one color will still be selectable from the "list colors"
option).