Adjust toolbar spacing for Qt 5.8.
It seems like some of the defaults have changed since 5.2.

I think stateLabel might have been already somewhat broken in older releases
(the colon was cut off?), but it's even worse now on Windows.

(I originally adjusted some of the spacing by setting margin in a stylesheet on
appropriate widgets, but on Windows: setting margin-left on a QSpinBox made the
up/down buttons flicker on hover, and setting margin-right on a QComboBox made
the text of the combo options shift left.)