Don't save the image name in project files.
Stop loading it from older project files as well - we don't actually use it in
projects.  The original intent was to maybe expose it somewhere for
informational purposes, or at least to have it around for debugging, but users
probably aren't aware that the filename is being stored, and we've never
actually needed it, so let's get rid of it.