Warn the user when we fail to write to a locked pdf.
Acroread (and maybe others?) lock any pdfs they're currently viewing, in which
case the call to begin the painter that draws on the printer that writes to the
pdf fails.  That's unfortunate since we could well be the one who opened the
viewer for the user to view a previous version of their pattern, but we seem to
be stuck with it.